[3/n] Lightweight Ray AIR API refactor
