+++ doc/source/data/batch_inference.rst
@@ -423,15 +423,15 @@ Suppose your cluster has 4 nodes, each with 16 CPUs. To limit to at most
 Using models from Ray Train
 ---------------------------
 
-Models that have been trained with :ref:`Ray Train <train-docs>` can then be used for batch inference with :ref:`Ray Data <data>` via the :class:`Checkpoint <ray.air.checkpoint.Checkpoint>` that is returned by :ref:`Ray Train <train-docs>`.
+Models that have been trained with :ref:`Ray Train <train-docs>` can then be used for batch inference with :ref:`Ray Data <data>` via the :class:`Checkpoint <ray.train.Checkpoint>` that is returned by :ref:`Ray Train <train-docs>`.
 
 **Step 1:** Train a model with :ref:`Ray Train <train-docs>`.
 
 .. testcode::
 
     import ray
+    from ray.train import ScalingConfig
     from ray.train.xgboost import XGBoostTrainer
-    from ray.air.config import ScalingConfig
 
     dataset = ray.data.read_csv("s3://anonymous@air-example-data/breast_cancer.csv")
     train_dataset, valid_dataset = dataset.train_test_split(test_size=0.3)
@@ -456,13 +456,13 @@ Models that have been trained with :ref:`Ray Train <train-docs>` can then be use
 
     ...
 
-**Step 2:** Extract the :class:`Checkpoint <ray.air.checkpoint.Checkpoint>` from the training :class:`Result <ray.air.Result>`.
+**Step 2:** Extract the :class:`Checkpoint <ray.train.Checkpoint>` from the training :class:`Result <ray.train.Result>`.
 
 .. testcode::
 
     checkpoint = result.checkpoint
 
-**Step 3:** Use Ray Data for batch inference. To load in the model from the :class:`Checkpoint <ray.air.checkpoint.Checkpoint>` inside the Python class, use one of the framework-specific Checkpoint classes.
+**Step 3:** Use Ray Data for batch inference. To load in the model from the :class:`Checkpoint <ray.train.Checkpoint>` inside the Python class, use one of the framework-specific Checkpoint classes.
 
 In this case, we use the :class:`XGBoostCheckpoint <ray.train.xgboost.XGBoostCheckpoint>` to load the model.
 
@@ -475,7 +475,7 @@ The rest of the logic looks the same as in the `Quickstart <#quickstart>`_.
     import numpy as np
     import xgboost
 
-    from ray.air import Checkpoint
+    from ray.train import Checkpoint
     from ray.train.xgboost import XGBoostCheckpoint
 
     test_dataset = valid_dataset.drop_columns(["target"])
