+++ doc/source/train/user-guides/persistent-storage.rst
@@ -25,7 +25,7 @@ The properties of the run configuration are :ref:`not tunable <tune-search-space
 
 .. seealso::
 
-    See the :class:`~ray.air.RunConfig` API reference.
+    See the :class:`~ray.train.RunConfig` API reference.
 
     See :ref:`tune-storage-options` for storage configuration examples (related to ``storage_path``).
