+++ doc/source/tune/api/execution.rst
@@ -29,7 +29,7 @@ Tuner Configuration
 
 .. seealso::
 
-    The `Tuner` constructor also takes in a :class:`air.RunConfig <ray.air.RunConfig>`.
+    The `Tuner` constructor also takes in a :class:`air.RunConfig <ray.train.RunConfig>`.
 
 Restoring a Tuner
 ~~~~~~~~~~~~~~~~~
