+++ doc/source/tune/tutorials/tune-fault-tolerance.rst
@@ -73,7 +73,7 @@ The actual training loop is just for demonstration purposes: the important detai
     :end-before: __ft_initial_run_end__
 
 The results and checkpoints of the experiment are saved to ``~/ray_results/tune_fault_tolerance_guide``,
-as configured by :class:`~ray.air.config.RunConfig`.
+as configured by :class:`~ray.train.RunConfig`.
 If the experiment has been interrupted due to one of the reasons listed above, use this path to resume:
 
 .. literalinclude:: /tune/doc_code/fault_tolerance.py
@@ -189,10 +189,10 @@ Trial-level fault tolerance deals with individual trial failures in the cluster,
 - Ephemeral network connection issues.
 - Nodes running out of memory or out of disk space.
 
-Ray Tune provides a way to configure failure handling of individual trials with the :class:`~ray.air.config.FailureConfig`.
+Ray Tune provides a way to configure failure handling of individual trials with the :class:`~ray.train.FailureConfig`.
 
 Assuming that we're using the ``trainable`` from the previous example that implements
-trial checkpoint saving and loading, here is how to configure :class:`~ray.air.config.FailureConfig`:
+trial checkpoint saving and loading, here is how to configure :class:`~ray.train.FailureConfig`:
 
 .. literalinclude:: /tune/doc_code/fault_tolerance.py
     :language: python
