+++ doc/source/tune/tutorials/tune-resources.rst
@@ -19,7 +19,7 @@ of CPUs (cores) on your machine.
     results = tuner.fit()
 
 You can override this per trial resources with :func:`tune.with_resources <ray.tune.with_resources>`. Here you can
-specify your resource requests using either a dictionary, a :class:`~ray.air.config.ScalingConfig`, or a
+specify your resource requests using either a dictionary, a :class:`~ray.train.ScalingConfig`, or a
 :class:`PlacementGroupFactory <ray.tune.execution.placement_groups.PlacementGroupFactory>`
 object. In any case, Ray Tune will try to start a placement group for each trial.
 
@@ -66,7 +66,7 @@ Even if the trial cannot be scheduled right now, Ray Tune will still try to star
 :ref:`autoscaling behavior <cluster-index>` if you're using the Ray cluster launcher.
 
 .. warning::
-    ``tune.with_resources`` cannot be used with :ref:`Ray Train Trainers <train-docs>`. If you are passing a Trainer to a Tuner, specify the resource requirements in the Trainer instance using :class:`~ray.air.config.ScalingConfig`. The general principles outlined below still apply.
+    ``tune.with_resources`` cannot be used with :ref:`Ray Train Trainers <train-docs>`. If you are passing a Trainer to a Tuner, specify the resource requirements in the Trainer instance using :class:`~ray.train.ScalingConfig`. The general principles outlined below still apply.
 
 It is also possible to specify memory (``"memory"``, in bytes) and custom resource requirements.
